chore(reviewrequest): reorganize components
