+++ src/layouts/ReviewRequest/components/ReviewRequestModal/ReviewRequestModal.stories.tsx
@@ -3,7 +3,6 @@ import { Story, ComponentMeta } from "@storybook/react"
 
 import { MOCK_ITEMS, MOCK_ADMINS } from "mocks/constants"
 
-import { EditedItemProps } from "./RequestOverview"
 import {
   ReviewRequestModal,
   ReviewRequestModalProps,
